form with name and amount fields rendered on frontend
