+++ client/components/groceryTable.jsx
@@ -15,7 +15,7 @@ function GroceryTable(props) {
     />
   ));
   return (
-    <table className='table table-striped table-bordered col-12'>
+    <table className='table table-striped table-bordered order-2 order-md-1 col-12 col-md-9'>
       <thead className='thead-light'>
         <tr>
           <th>Item Name</th>
